+++ backend/package.json
@@ -7,6 +7,7 @@
   },
   "devDependencies": {
     "@types/bun": "latest",
+    "@types/jsonwebtoken": "^9.0.9",
     "drizzle-kit": "^0.31.1"
   },
   "peerDependencies": {
@@ -15,6 +16,7 @@
   "dependencies": {
     "dotenv": "^16.5.0",
     "drizzle-orm": "^0.44.2",
-    "jsonwebtoken": "^9.0.2"
+    "jsonwebtoken": "^9.0.2",
+    "tiny-invariant": "^1.3.3"
   }
 }
