+++ backend/src/entities/boss.ts
@@ -11,29 +11,28 @@ import {
 } from "../../../core/game";
 import {
   addComponent,
-  addUpdateCallback,
   createEntity,
   destroyEntity,
   runQuery,
 } from "../ecsProvider";
+import { addUpdateCallback } from "../update";
 import { createProjectile } from "./projectile";
 
 export const BOSS_COMPONENT_DEF = {
   type: "boss",
-  detectionRange: 300, // Detection range for players
-  shootCooldown: 60, // Frames between shots (increased from 30)
-  currentCooldown: 0, // Current cooldown counter
-  patternCooldown: 300, // Frames between special attack patterns (increased from 180)
-  currentPatternCooldown: 0, // Current pattern cooldown counter
-  currentPattern: 0, // Current attack pattern (0: none, 1: circle, 2: spiral, 3: random burst)
-  patternStep: 0, // Current step in the pattern
-  patternDuration: 90, // How long pattern lasts (reduced from 120)
-  // New movement properties
-  moveDirection: { x: 0, y: 0 }, // Current movement direction
-  moveDuration: 0, // Frames remaining in current direction
-  moveSpeed: 1.5, // Reduced base movement speed (was 4)
-  minMoveDuration: 120, // Increased minimum duration (was 30)
-  maxMoveDuration: 240, // Increased maximum duration (was 90)
+  detectionRange: 300,
+  shootCooldown: 60,
+  currentCooldown: 0,
+  patternCooldown: 300,
+  currentPatternCooldown: 0,
+  currentPattern: 0,
+  patternStep: 0,
+  patternDuration: 90,
+  moveDirection: { x: 0, y: 0 },
+  moveDuration: 0,
+  moveSpeed: 1.5,
+  minMoveDuration: 120,
+  maxMoveDuration: 240,
 };
 
 export const createBossEntity = (entity: Entity) => {
@@ -58,7 +57,7 @@ const circleAttack = (x: number, y: number, bulletCount: number) => {
     const angle = (i / bulletCount) * Math.PI * 2;
     const velocityX = Math.cos(angle) * 5;
     const velocityY = Math.sin(angle) * 5;
-    createProjectile(createEntity(), x, y, velocityX, velocityY, "boss", 15); // Circle attack does more damage
+    createProjectile(createEntity(), x, y, velocityX, velocityY, "boss", 15);
   }
 };
 
@@ -66,7 +65,7 @@ const spiralAttack = (x: number, y: number, step: number) => {
   const angle = (step / 20) * Math.PI * 2;
   const velocityX = Math.cos(angle) * 5;
   const velocityY = Math.sin(angle) * 5;
-  createProjectile(createEntity(), x, y, velocityX, velocityY, "boss", 20); // Spiral attack does the most damage
+  createProjectile(createEntity(), x, y, velocityX, velocityY, "boss", 20);
 };
 
 const randomBurst = (x: number, y: number, bulletCount: number) => {
@@ -74,7 +73,7 @@ const randomBurst = (x: number, y: number, bulletCount: number) => {
     const angle = Math.random() * Math.PI * 2;
     const velocityX = Math.cos(angle) * 5;
     const velocityY = Math.sin(angle) * 5;
-    createProjectile(createEntity(), x, y, velocityX, velocityY, "boss", 10); // Random burst does less damage
+    createProjectile(createEntity(), x, y, velocityX, velocityY, "boss", 10);
   }
 };
 
@@ -82,9 +81,7 @@ addUpdateCallback(() => {
   runQuery(
     [POSITION_COMPONENT_DEF, BOSS_COMPONENT_DEF],
     (_entity, [position, boss]) => {
-      // Update movement direction if needed
       if (boss.moveDuration <= 0) {
-        // Pick a new random direction
         const angle = Math.random() * Math.PI * 2;
         boss.moveDirection.x = Math.cos(angle);
         boss.moveDirection.y = Math.sin(angle);
@@ -94,27 +91,22 @@ addUpdateCallback(() => {
         );
       }
 
-      // Apply movement in current direction
       position.x += boss.moveDirection.x * boss.moveSpeed;
       position.y += boss.moveDirection.y * boss.moveSpeed;
 
-      // Add very slight random variation to movement
-      const randomX = (Math.random() * 2 - 1) * 0.2; // Reduced from 0.5
-      const randomY = (Math.random() * 2 - 1) * 0.2; // Reduced from 0.5
+      const randomX = (Math.random() * 2 - 1) * 0.2;
+      const randomY = (Math.random() * 2 - 1) * 0.2;
       position.x += randomX;
       position.y += randomY;
 
-      // Less frequent dodges
       if (Math.random() < 0.01) {
-        // Reduced from 0.02
         const dodgeAngle = Math.random() * Math.PI * 2;
-        position.x += Math.cos(dodgeAngle) * 8; // Reduced from 12
-        position.y += Math.sin(dodgeAngle) * 8; // Reduced from 12
+        position.x += Math.cos(dodgeAngle) * 8;
+        position.y += Math.sin(dodgeAngle) * 8;
       }
 
       boss.moveDuration--;
 
-      // Rest of the existing boss logic
       if (boss.currentCooldown > 0) {
         boss.currentCooldown--;
       }
@@ -126,43 +118,34 @@ addUpdateCallback(() => {
         if (boss.currentPatternCooldown > 0) {
           boss.currentPatternCooldown--;
         } else {
-          boss.currentPattern = Math.floor(Math.random() * 3) + 1; // 1-3
+          boss.currentPattern = Math.floor(Math.random() * 3) + 1;
           boss.patternStep = 0;
           boss.currentPatternCooldown = boss.patternDuration;
         }
       } else {
-        // Pattern is active
         if (boss.currentPatternCooldown > 0) {
           boss.currentPatternCooldown--;
 
-          // Execute the current pattern
           switch (boss.currentPattern) {
             case 1:
-              // Circle pattern - shoot bullets in a circle
               if (boss.patternStep % 20 === 0) {
-                // Every 20 frames (increased from 10)
-                circleAttack(bossX, bossY, 6); // 6 bullets in a circle (reduced from 12)
+                circleAttack(bossX, bossY, 6);
               }
               break;
             case 2:
-              // Spiral pattern - shoot bullets in a spiral
               if (boss.patternStep % 5 === 0) {
-                // Every 5 frames
                 spiralAttack(bossX, bossY, boss.patternStep);
               }
               break;
             case 3:
-              // Random burst pattern - shoot bullets in random directions
               if (boss.patternStep % 15 === 0) {
-                // Every 15 frames
-                randomBurst(bossX, bossY, 3); // 3 bullets per burst (reduced from 5)
+                randomBurst(bossX, bossY, 3);
               }
               break;
           }
 
           boss.patternStep++;
         } else {
-          // Pattern is complete, reset
           boss.currentPattern = 0;
           boss.currentPatternCooldown = boss.patternCooldown;
         }
@@ -171,7 +154,6 @@ addUpdateCallback(() => {
   );
 });
 
-// Add collision handling for boss-projectile collisions
 addUpdateCallback(() => {
   runQuery(
     [
@@ -181,30 +163,24 @@ addUpdateCallback(() => {
       BOSS_COMPONENT_DEF,
     ],
     (bossEntity, [bossPos, bossCollider, bossHealth, _boss]) => {
-      // Get all projectiles that could be colliding with the boss
       runQuery(
         [POSITION_COMPONENT_DEF, PROJECTILE_COMPONENT_DEF],
         (projectileEntity, [projectilePos, projectile]) => {
-          // Skip projectiles created by the boss
           if (projectile.source === "boss") return;
 
-          // Check for collision using circle collision
           const dx = bossPos.x - projectilePos.x;
           const dy = bossPos.y - projectilePos.y;
           const distanceSquared = dx * dx + dy * dy;
           const combinedRadius = bossCollider.radius + projectile.radius;
 
           if (distanceSquared <= combinedRadius * combinedRadius) {
-            // Boss takes damage
             bossHealth.currentHealth = Math.max(
               0,
               bossHealth.currentHealth - projectile.damage,
             );
 
-            // Destroy the projectile
             destroyEntity(projectileEntity);
 
-            // If boss health reaches 0, destroy the boss entity
             if (bossHealth.currentHealth <= 0) {
               destroyEntity(bossEntity);
             }
